Extract file utils to separate module
