+++ benchmarks/evaluate.py
@@ -4,7 +4,7 @@ from utils import get_result, download_dataset, get_frame_sampling_permutations
 import os
 from winnow.utils.config import resolve_config
 import click
-from winnow.utils import scan_videos
+from winnow.utils.files import scan_videos
 import subprocess
 import shlex
 import numpy as np
