+++ extract_exif.py
@@ -6,11 +6,11 @@ from db import Database
 from winnow.storage.db_result_storage import DBResultStorage
 from winnow.storage.repr_utils import path_resolver
 from winnow.utils import (
-    scan_videos,
     extract_from_list_of_videos,
     convert_to_df,
     parse_and_filter_metadata_df,
 )
+from winnow.utils.files import scan_videos
 from winnow.utils.config import resolve_config
