+++ winnow/utils/utils.py
@@ -1,8 +1,6 @@
 import hashlib
 import json
 import os
-from glob import glob
-from pathlib import Path
 
 import cv2
 import numpy as np
@@ -10,64 +8,10 @@ from joblib import load
 
 from winnow.storage.repr_key import ReprKey
 from winnow.storage.repr_utils import path_resolver
+from winnow.utils.files import get_hash
 
 DEFAULT_DIRECTORY = os.path.join(os.path.dirname(__file__), "models")
 GRAY_ESTIMATION_MODEL = os.path.join(DEFAULT_DIRECTORY, "gb_gray_model.joblib")
-
-
-def create_directory(directories, root_dir, alias):
-
-    for r in directories:
-        try:
-            os.makedirs(os.path.abspath(os.path.join(root_dir, alias, r)))
-        except Exception as e:
-            print(e)
-
-
-def filter_extensions(files, extensions):
-    extensions = [f".{ext}" for ext in extensions]
-    return [x for x in files if Path(x).suffix in extensions]
-
-
-def scan_videos(path, wildcard, extensions=[]):
-    """Scans a directory for a given wildcard
-
-    Args:
-        path (String): Root path of the directory to be scanned
-        wildcard (String): Wild card related to the files being searched
-        (eg. ** for video files or **_vgg_features.npy for extracted features
-        files) extensions (list, optional): Filter files by giving a list of
-        supported file extensions (eg a list of video extensions).
-        Defaults to [].
-
-    Returns:
-        List[String]: A list of file paths
-    """
-
-    files = glob(os.path.join(path, wildcard), recursive=True)
-    files = [x for x in files if os.path.isfile(x)]
-    if len(extensions) > 0:
-        files = filter_extensions(files, extensions)
-
-    return files
-
-
-def scan_videos_from_txt(fp, extensions=[]):
-
-    files = list(open(fp, encoding="utf-8").read().splitlines())
-    files = [x for x in files if os.path.isfile(x)]
-    if len(extensions) > 0:
-        files = filter_extensions(files, extensions)
-    return files
-
-
-def create_video_list(videos_to_be_processed, fp):
-
-    with open(fp, "w", encoding="utf-8") as f:
-        for item in videos_to_be_processed:
-            f.write("%s\n" % item)
-
-    return os.path.abspath(fp)
 
 
 def filter_results(thr, distances, indices):
@@ -149,19 +93,6 @@ def extract_additional_info(reps, repr_key):
     std_sum = np.std(intra_sum)
 
     return (shape[0], mean_act, std_sum, max_dif, grays_avg, grays_std, grays_max)
-
-
-def get_hash(fp, buffer_size=65536):
-
-    sha256 = hashlib.sha256()
-    with open(fp, "rb") as f:
-        while True:
-            data = f.read(buffer_size)
-            if not data:
-                break
-            sha256.update(data)
-
-    return sha256.hexdigest()
 
 
 def get_config_tag(config):
